+++ gym_rarl/__init__.py
@@ -1,0 +1,6 @@
+# from gym.envs.registration import register
+#
+# register(
+#     id='Adv-Acrobot-v1',
+#     entry_point='gym_rarl.envs:AdversarialAcrobotEnv',
+# )
