+++ gym_rarl/envs/__init__.py
@@ -1,0 +1,2 @@
+# from gym_rarl.envs.rarl_acrobot import AdversarialAcrobotEnv
+# from gym_rarl.envs.rarl_env import MainRarlEnv, AdversarialRarlEnv
