+++ setup.py
@@ -1,0 +1,6 @@
+# from setuptools import setup
+#
+# setup(name='gym_rarl',
+#       version='0.0.1',
+#       install_requires=['gym']  # And any other dependencies foo needs
+#       )
